Fix all clippy warnings for Rust 1.94+ stable
- Replace sort_by with sort_by_key (clippy::sort_by_key_reverse)
- Collapse if-in-match arms (clippy::collapsible_match)
- Remove unnecessary same-type casts (u8->u8, i128->i128, usize->usize)
- Use ? operator where applicable (clippy::question_mark)

+++ src/interpreter/builtins/temporal/plain_time.rs
@@ -229,7 +229,7 @@ impl Interpreter {
                     };
 
                     let is_neg = rounded < 0;
-                    let abs_ns = if is_neg { -rounded } else { rounded } as i128;
+                    let abs_ns = if is_neg { -rounded } else { rounded };
                     let sign_f = if is_neg {
                         -1.0
                     } else if abs_ns == 0 {
